// src/app/model/base.ts
export enum Dice {
  D4 = 'D4',
  D6 = 'D6',
  D8 = 'D8',
  D10 = 'D10',
  D12 = 'D12',
  D20 = 'D20'
};

export enum Language {
  ABYSSAL = 'Abyssal',
  AURAN = 'Auran',
  COMMON = 'Common',
  DEEP_SPEECH = 'Deep Speech',
  DRACONIC = 'Draconic',
  DRUIDIC = 'Druidic',
  DWARVISH = 'Dwarvish',
  ELEMENTAL = 'Elemental',
  ELVISH = 'Elvish',
  GIANT = 'Giant',
  GNOMISH = 'Gnomish',
  GOBLIN = 'Goblin',
  HALFLING = 'Halfling',
  INFERNAL = 'Infernal',
  ORCISH = 'Orcish',
  PRIMORDIAL = 'Primordial',
  SYLVAIN = 'Sylvain',
  UNDERCOMMON = 'Undercommon'
};

// export enum HairColor {
//   BALD = 'bald',
//   BLACK = 'black',
//   DARK_BROWN = 'dark brown',
//   MEDIUM_BROWN = 'medium brown',
//   LIGHT_BROWN = 'light brown',
//   CHESTNUT = 'chestnut',
//   AUBURN = 'auburn',
//   RED = 'red',
//   GINGER = 'ginger',
//   COPPER = 'copper',
//   STRAWBERRY = 'strawberry',
//   LIGHT_BLOND = 'light blond',
//   MEDIUM_BLOND = 'medium blond',
//   GREY = 'grey',
//   WHITE = 'white'
// };

// export enum EyeColor {
//   AMBER = 'amber',
//   BLUE = 'blue',
//   BROWN = 'brown',
//   GREY = 'grey',
//   GREEN = 'green',
//   HAZEL = 'hazel',
//   RED = 'red',
//   VIOLET = 'violet'
// };

export enum Alignment {
  CHAOTIC_GOOD = 'Chaotic good',
  CHAOTIC_NEUTRAL = 'Chaotic neutral',
  CHAOTIC_EVIL = 'Chaotic evil',
  NEUTRAL_GOOD = 'Neutral good',
  TRUE_NEUTRAL = 'True neutral',
  NEUTRAL_EVIL = 'Neutral evil',
  LAWFUL_GOOD = 'Lawful good',
  LAWFUL_NEUTRAL = 'Lawful neutral',
  LAWFUL_EVIL = 'Lawful evil'
};